Añade constructor de informe #3
